Finished The Short Personal Project
I do not know if I want to add much more to it, but I'm keeping it at is. I hope to spend one afternoon encoding my favorite texts with the Caesar Cipher or sending letters to my friends with secret messages that I won't have a tough time decoding. I hope this brings joy to people. I think it's enjoyable.